requirements.txt
requests>=2.31.0
urllib3>=2.0.0
PyYAML>=6.0.1
scapy>=2.5.0
openai>=1.30.0
anthropic>=0.28.0
rich>=13.7.0
click>=8.1.7
python-dateutil>=2.9.0
tabulate>=0.9.0
